Moved legacy estimate to its own module for fallback ops/backwards compat
