+++ Cargo.toml
@@ -22,10 +22,7 @@ python = ["dep:pyo3"]
 [dependencies]
 regex = "1.8.4"
 lazy_static = "1.4.0"
-rayon = "1.9.0"
 pyo3 = { version = "0.25.1", optional = true }
-serde_json = "1.0.68"
-hashbrown = "0.16"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
