--- src/syllables.rs
@@ -1,1 +1,0 @@
-// Placeholder - this module is superseded by src/cmudict.rs
